pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Arina Zubova <arina777pototskaya@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
django-bootstrap4 = "^23.1"
python-dotenv = "^1.0.0"
flake8 = "^6.0.0"
dj-database-url = "^1.3.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
